Fixed excessive xp generation.
Awarding 5 xp for every 100 blocks the player ever dug or placed, is shurly not what you intended. 
For example if the player dug 1000 stone that will result in 50 xp. At 1100, 55 xp and so on. 
Mining with an bronze pick let's you easily dig 10,000 stone in one afternoon (this is what i did) resulting in 25,250 xp.

The changes will award the player with an average of 2.5 xp per 100 blocks.